fix: change samples to auto create connect user if one is not found
For the sake of simplicity, I changed the LoginMenu code to auto-create
a connect user. In a real title, it would make sense to ask/confirm that a new connect
user is supposed to be made, in case an error was made and the user was
expecting it to match with an existing account.