EOSConfig to ScriptableObject, enable Windows Standalone build to create Platform
EOSConfig turns to Scriptable Object mean the config now can be created and placed anywhere in the project, not just Streaming Asset Folder, and do not need to be found by path. Just create one and put in the EOSManager.

Also add a fix so that Window Standalone Build can now create Platform.