fix(config,steam): compile errors on platforms that don't support steamworks
This commit fixes an error happening in the Steam config tool.

This is likely an overkill with the refactor however.

The easier fix is to just add the whole line of platform define for DISABLESTEAMWORKS into EpicOnlineServicesConfigEditor.cs
This is efficient if a small number of scripts uses steamworks.
If it grows the same define would need to be added all the scripts