Update: EOSLobbyManager, EOSTransport, EOSEACLobbyManager: Added additional Voice Channel related functions
Added new MuteLocalMember and DeafenLocalMember functions which do not rely on CurrentLobby.Members List. The existing ToggleMute function relies on the CurrentLobby.Members list to be populated and will not work before that update is received. MuteLocalMember and DeafenLocalMember can be execute as soon as the Lobby is created.

Added new MuteMember function that will mute a Target ProductUserId. It is similar to the existing ToggleMute function just that this one isn't a toggle and will set Mute status based of the specified value.